+++ DeepINN/geometry/__init__.py
@@ -1,5 +1,1 @@
-# Make the subdirectory callable
-from . import domains
-#from . import conditions
-from . import samplers
-from . import spaces
+pass
