+++ DeepINN/geometry/points.py
@@ -1,0 +1,8 @@
+"""
+Contains a class that handles the storage of all created data points.
+"""
+
+import torch
+import numpy as np
+from .space import Space
+
