+++ DeepINN/geometry_legacy/__init__.py
@@ -1,0 +1,5 @@
+# Make the subdirectory callable
+from . import domains
+#from . import conditions
+from . import samplers
+from . import spaces
