+++ DeepINN/geometry_legacy/conditions/deeponet_condition.py
