+++ DeepINN/geometry_legacy/domains/domain1D/interval.py
