+++ DeepINN/geometry_legacy/domains/domain2D/triangle.py
