+++ DeepINN/geometry_legacy/domains/domain3D/trimesh_polyhedron.py
